blog/requirements.txt
Django==4.1.1
amqp                       5.1.1
asgiref                    3.5.2
backports.zoneinfo         0.2.1
billiard                   3.6.4.0
botocore                   1.29.27
celery                     5.2.7
certifi                    2022.9.24
charset-normalizer         2.1.1
click                      8.1.3
click-didyoumean           0.3.0
click-plugins              1.1.1
click-repl                 0.2.0
colorama                   0.4.5
crispy-bootstrap5          0.7
Django                     3.2.16
django-appconf             1.0.5
django-celery              3.1.17
django-ckeditor            6.5.1
django-crispy-forms        1.14.0
django-dynamic-admin-forms 0.1.8
django-encode              1.0.4
django-jazzmin             2.5.0
django-js-asset            2.0.0
django-queued-storage      0.8.0
django-richtextfield       1.6.1
django-smart-selects       1.6.0
docopt                     0.6.2
idna                       3.4
jmespath                   1.0.1
kombu                      5.2.4
Pillow                     9.2.0
pip                        22.3.1
pipreqs                    0.4.11
prompt-toolkit             3.0.31
python-dateutil            2.8.2
pytz                       2022.2.1
requests                   2.28.1
s3transfer                 0.6.0
setuptools                 56.0.0
six                        1.16.0
sqlparse                   0.4.3
tzdata                     2022.4
urllib3                    1.26.12
vine                       5.0.0
wcwidth                    0.2.5
yarg                       0.1.9